♻️ refactor: index.ts 파일 추가 및 폴더 구조 개선
- events/ 및 commands/ 폴더에 index.ts 파일을 추가하여 모듈을 경로 간략화 형식으로 import 할 수 있도록 개선
- ex. `import {} from './events/index.ts';`
