+++ src/events/index.ts
@@ -1,0 +1,4 @@
+import InteractionCreate from './interactionCreate.ts';
+import Ready from './ready.ts';
+
+export { InteractionCreate, Ready };
