+++ src/types/index.ts
@@ -1,0 +1,4 @@
+import { Command } from './Command.ts';
+import { ExtendedClient } from './ExtendedClient.ts';
+
+export { Command, ExtendedClient };
